Refactor database models and schemas
