+++ app/database/models/measurement.py
@@ -1,7 +1,7 @@
 from sqlalchemy import CheckConstraint, Integer, String, SmallInteger
 from sqlalchemy.orm import Mapped, mapped_column
 from typing import Optional
-from app.database.models.Base import Base
+from app.database.models.base import Base
 
 
 class Measurement(Base):
