+++ app/service/calculator_service.py
@@ -1,7 +1,7 @@
 import os
 from sqlalchemy import create_engine
 from sqlalchemy.orm import Session
-from app.models.number import Number
+from app.database.models.number import Number
 
 
 class CalculatorService:
