+++ requirements.txt
@@ -2,6 +2,7 @@ fastapi==0.108.0
 uvicorn==0.25.0
 python-dotenv
 flake8
+psycopg2
 
 ### RabbitMQ ###
 pika
